Move classes and typealiases to separate files
