+++ EssentialDevelopper/EssentialDeveloperFramework/Feed API/Helpers/HTTPURLResponse+isOK.swift
@@ -1,0 +1,13 @@
+//
+//  HTTPURLResponse+isOK.swift
+//  EssentialFeed
+//
+//  Created by Andre Kvashuk on 6/19/22.
+//  Copyright © 2022 Andre Kvashuk. All rights reserved.
+//
+
+import Foundation
+
+extension HTTPURLResponse {
+    var isOK: Bool { statusCode == 200 }
+}
