+++ EssentialDevelopper/EssentialDeveloperFramework/Feed API/Helpers/SharedTypaliases.swift
@@ -1,0 +1,12 @@
+//
+//  SharedTypaliases.swift
+//  EssentialFeed
+//
+//  Created by Andre Kvashuk on 6/19/22.
+//  Copyright © 2022 Andre Kvashuk. All rights reserved.
+//
+
+import Foundation
+
+public typealias Closure<T> = (T) -> Void
+public typealias VoidClosure = () -> Void
